+++ towerGame/src/save/SaveFile.java
@@ -4,6 +4,7 @@ import java.awt.Color;
 import java.awt.Rectangle;
 import java.awt.image.BufferedImage;
 import java.io.*;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -56,7 +57,7 @@ public class SaveFile {
 			}else {
 				sd4.setObject(level.player.x,"playerX");
 				sd4.setObject(level.player.y,"playerY");
-				sd4.setObject(level.player.health,"playerHealth");
+				sd4.setObject(level.player.health.doubleValue(),"playerHealth");
 				sd4.setObject(level.player.mana,"playerMana");
 				sd4.setObject(level.player.armor,"playerArmor");
 				sd4.setObject(level.player.weapon,"playerWeapon");
@@ -104,7 +105,7 @@ public class SaveFile {
 				if(!level.inLevelEditor) {
 					level.player.x=(double)gs.attr.getObjectDefault("playerX",4.0D);
 					level.player.y=(double)gs.attr.getObjectDefault("playerY",6.0D);
-					level.player.health=(double)gs.attr.getObjectDefault("playerHealth",10.0D);
+					level.player.health=BigDecimal.valueOf((double)gs.attr.getObjectDefault("playerHealth",10.0D));
 					level.player.mana=(double)gs.attr.getObjectDefault("playerMana",15.0D);
 					level.player.armor=(double)gs.attr.getObjectDefault("playerArmor",0.0D);
 					level.player.setWeapon((int)gs.attr.getObjectDefault("playerWeapon",1));
@@ -166,7 +167,7 @@ public class SaveFile {
 					SerializedData player = (SerializedData) sd2.getObject("player");
 					level.player.x=(double)player.getObjectDefault("playerX",level.playerStartX);
 					level.player.y=(double)player.getObjectDefault("playerY",level.playerStartY);
-					level.player.health=(double)player.getObjectDefault("playerHealth",10.0D);
+					level.player.health=BigDecimal.valueOf((double)player.getObjectDefault("playerHealth",10.0D));
 					level.player.mana=(double)player.getObjectDefault("playerMana",15.0D);
 					level.player.armor=(double)player.getObjectDefault("playerArmor",0.0D);
 					level.player.setWeapon((int)player.getObjectDefault("playerWeapon",1));
